Removed shutdown and hello_word from application and they are now only defined in the test. Also disabled debug mode for the rest server in __main__.py.
