+++ src/__main__.py
@@ -3,4 +3,4 @@
 from rest_api.application import app, hostURL, hostPort
 
 if __name__ == "__main__":
-    app.run(host=hostURL, port=hostPort, debug=True)
+    app.run(host=hostURL, port=hostPort, debug=False)
